Updated param table and header

+++ apps/fac/fsw/src/fac_tbldefs.h
@@ -404,7 +404,7 @@ typedef struct
 	 * @boolean
 	 * @group FW Attitude Control
 	 */
-	float FW_W_EN;
+	int32 FW_W_EN;
 
 	/**
 	 * Wheel steering rate proportional gain
